% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_diff.R
\name{estimate_joint_prior}
\alias{estimate_joint_prior}
\title{Estimate the joint for a mark}
\usage{
estimate_joint_prior(input, n_case = 12, iter = 2000, use_vb = TRUE,
  use_subsamp = TRUE, n_subsamp = 1000)
}
\arguments{
\item{input}{character - path to an mark input file}

\item{n_case}{integer - the number of columns (after the id column) that are cases. The remainder are considered controls.}

\item{iter}{integer - number of posterior samples to draw}

\item{use_vb}{logical - use a variational approximation}
}
\value{
a stanfit object containing the marginal joint prior on case_a, case_b, ctrl_a, and ctrl_b
}
\description{
Fits a stan model to estimate the joint distribution on mark probability between two groups.
}
\details{
This function takes in a file of marks. The input should be a tsv, with the first column giving a region id and the remainder giving peak calls for each subject. Non-zero integer counts are converted to TRUE, while 0's are converted to FALSE.
    The package fits a joint beta distribution on the peak probabilities in the two groups.
}
