% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_diff.R
\name{vb_approx}
\alias{vb_approx}
\title{Variational approximation of mark differences}
\usage{
vb_approx(mark_dat, n_ctrl, n_case, iter)
}
\description{
This functional uses \code{rstan::vb} to run a variational approximation on the probability
}
